feat: Implement color palette extraction and refactor MainActivity
- Implemented color palette extraction from images using the Palette API to dynamically determine dominant colors.
- Added `colorPalet.kt` with functions to extract dominant colors from bitmaps, handling different bitmap configurations.
- Refactored `MainActivity` by renaming it and moving it to `presentation` package.
- Added `YouTubePlayerManager` and injected it and Application instance into `HomeViewModel` via Koin.
