+++ app/src/main/java/com/example/seekhoandoridassignment/di/appModule.kt
@@ -1,10 +1,14 @@
 package com.example.seekhoandoridassignment.di
 
+import android.app.Application
 import com.example.seekhoandoridassignment.BuildConfig.BASE_URL
 import com.example.seekhoandoridassignment.data.network.AnimeRepositoryImp
 import com.example.seekhoandoridassignment.data.network.ApiService
 import com.example.seekhoandoridassignment.domain.repository.AnimeRepository
 import com.example.seekhoandoridassignment.presntation.viewmodels.HomeViewModel
+import com.example.seekhoandoridassignment.uitl.YouTubePlayerManager
+import org.koin.android.ext.koin.androidApplication
+import org.koin.android.ext.koin.androidContext
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
 import retrofit2.Retrofit
@@ -12,10 +16,12 @@ import retrofit2.converter.gson.GsonConverterFactory
 
 val  appModule = module {
  single { provideConverterFactory() }
+ single { androidApplication()}
  single { provideRetrofit(get()) }
  single { provideService(get()) }
  single { provideAnimeRepository(get()) }
- viewModel { HomeViewModel(get()) }
+  single { YouTubePlayerManager(get()) }
+ viewModel { HomeViewModel(get(),get()) }
 
 
 }
